fix declaration error

+++ src/obex_auth.c
@@ -91,11 +91,11 @@ int obex_auth_add_challenge (obex_t* handle,
 	return 0;
 }
 
-int obex_auth_unpack_response (obex_headerdata_t h,
-                               uint32_t size,
-                               /* out */ uint8_t digest[16],
-                               /* out */ uint8_t nonce[16],
-                               /* out */ uint8_t user[21])
+ssize_t obex_auth_unpack_response (obex_headerdata_t h,
+				   uint32_t size,
+				   /* out */ uint8_t digest[16],
+				   /* out */ uint8_t nonce[16],
+				   /* out */ uint8_t user[21])
 {
 	int len = 0;
 	uint32_t i = 0;
